+++ src/main/java/org/informatics/service/GoodsService.java
@@ -2,8 +2,10 @@ package org.informatics.service;
 
 import org.informatics.data.Goods;
 
+import java.math.BigDecimal;
+
 public interface GoodsService {
         boolean isExpired(Goods goods);
         long daysUntilExpiry(Goods goods);
-        double calculateSellingPrice(Goods goods);
+        BigDecimal calculateSellingPrice(Goods goods);
 }
