+++ src/main/java/org/informatics/service/ReceiptService.java
@@ -6,9 +6,10 @@ import org.informatics.data.Shop;
 
 
 import java.io.IOException;
+import java.math.BigDecimal;
 
 public interface ReceiptService  {
-    public double calculateTotalPrice(Receipt receipt);
+    public BigDecimal calculateTotalPrice(Receipt receipt);
     public void addGoods(Receipt receipt, Goods good);
     public void serializeObj(Receipt receipt, String filename) throws IOException;
     Receipt deserializeObj(String filename) throws IOException, ClassNotFoundException;
